requirements.txt
click
pyyaml
python-dotenv==1.2.1